16:05までの進捗
cooldownfpsとangleの固定化,wavと音楽の追加
