+++ toruyohpractice/Game1/Datas/MusicPlayer2.cs
@@ -28,8 +28,12 @@ namespace CommonPart
         //BGMIDを同じにしないでください。曲名を省略するとファイルパスからファイル名を取得し、それをそのまま曲名とする。
         public BGMdata[] bgmDatas = new BGMdata[] {
             new BGMdata("Content/barragetitle.wav",BGMID.title,100), // ループしない曲はループ情報を省略するだけ
-            new BGMdata("Content/Stage2_Boss.wav",BGMID.Stage2Boss,100,6479,232083), //ファイル名が曲名になる
+            new BGMdata("Content/Stage2_Boss.wav",BGMID.Stage2Boss,100,"蹂躙と幻想ノ狭間ニ",6479,232083), //ファイル名が曲名になる
             new BGMdata("Content/stage４ボス.wav",BGMID.Stage4Boss,100,"花弁",47340,207970),
+            new BGMdata("Content/stage1_way.wav",BGMID.Stage1onWay,100,"埃をかぶった風"),
+            new BGMdata("Content/iseki.wav",BGMID.Stage1Boss,100,"Metalworms"),
+            new BGMdata("Content/maten_game.wav",BGMID.Stage2onWay,100,"the Paradox City"),
+            new BGMdata("Content/Stage4_Journey_short.wav",BGMID.Stage4onWay,100,"雪解けの春"),
         };
