+++ toruyohpractice/Game1/Game1.csproj
@@ -79,9 +79,9 @@
     <Compile Include="Window\Window_AniDList.cs" />
     <Compile Include="Window\Window_Animation.cs" />
     <Compile Include="Window\Window_UnitType.cs" />
-    <Compile Include="Workers\Unit.cs" />
-    <Compile Include="Workers\Worker.cs" />
-    <Compile Include="Workers\WorkerManager.cs" />
+    <Compile Include="Window\Workers\Unit.cs" />
+    <Compile Include="Window\Workers\Worker.cs" />
+    <Compile Include="Window\Workers\WorkerManager.cs" />
     <Compile Include="XNA\AnimationAdvanced.cs" />
     <Compile Include="XNA\Inputmanager.cs" />
     <Compile Include="Game1.cs" />
@@ -184,10 +184,19 @@
     <Content Include="Content\Boss1_tail.png">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Content>
+    <Content Include="Content\gameover.wav">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </Content>
+    <Content Include="Content\iseki.wav">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </Content>
     <Content Include="Content\KMC.png">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Content>
     <Content Include="Content\leftside1.jpg">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </Content>
+    <Content Include="Content\maten_game.wav">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Content>
     <Content Include="Content\None.png">
@@ -211,7 +220,11 @@
     <Content Include="Content\scroll.png">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </Content>
-    <Content Include="Content\Stage2_Boss.wav">
+    <Content Include="Content\stage1_way.wav">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </Content>
+    <Content Include="Content\Stage2_Boss.wav" />
+    <Content Include="Content\Stage4_Journey_short.wav">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Content>
     <Content Include="Content\stage４ボス.wav">
